+++ CHANGELOG.md
@@ -4,5 +4,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+* Add Spin::fn for retrying assertions
 * Add DateParam::parse for parsing relative date strings from features
 * Project commenced